+++ src/Settings.ts
@@ -2,6 +2,7 @@ export class Settings {
 
     static NAMESPACE = "bsa-wfrp4e";
     public static SKILLS = "skills";
+    public static ALLOW_CRITICAL = "allowCritical";
 
     static init() {
         game["settings"].register(this.NAMESPACE, this.SKILLS, {
@@ -12,6 +13,14 @@ export class Settings {
             default: "",
             requiresReload: true,
             type: String,
+        });
+        game["settings"].register(this.NAMESPACE, this.ALLOW_CRITICAL, {
+            name: "Allow critical rolls",
+            hint: "Allow critical success and failures e.g. one roll can produce multiple fails or success",
+            scope: "world",
+            config: true,
+            default: false,
+            type: Boolean,
         });
     }