+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.1.0
+- add SkillTest
+- breaking change no dc you need to create a success
+- breaking change it can now produce multiple successes and fails with on single roll.
+- add setting to allow this (Allow critical rolls)
 # 1.0.2
 - fix for core module
 # 1.0.1